plot four game UI
