+++ src/com/ukan/plotfour/Controller.java
@@ -1,0 +1,26 @@
+package com.ukan.plotfour;
+
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Label;
+import javafx.scene.layout.GridPane;
+import javafx.scene.layout.Pane;
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class Controller implements Initializable {
+
+	@FXML
+	public GridPane rootGridPane;
+
+	@FXML
+	public Pane insertedDiscsPane;
+
+	@FXML
+	public Label playerNameLabel;
+
+	@Override
+	public void initialize(URL location, ResourceBundle resources) {
+
+	}
+}
