+++ src/com/ukan/plotfour/Main.java
@@ -1,0 +1,34 @@
+package com.ukan.plotfour;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.layout.GridPane;
+import javafx.stage.Stage;
+
+public class Main extends Application {
+
+    private  Controller controller;
+
+    @Override
+    public void start(Stage primaryStage) throws Exception{
+
+        FXMLLoader loader = new FXMLLoader(getClass().getResource("game.fxml"));
+        GridPane rootGridPane = loader.load();
+
+        controller = loader.getController();
+
+        Scene scene = new Scene(rootGridPane);
+
+        primaryStage.setScene(scene);
+        primaryStage.setTitle("Plot Four");
+        primaryStage.setResizable(false);
+        primaryStage.show();
+    }
+
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+}
